Implemented memory overrides for new/delete. Use the C++17 compliant __STDCPP_DEFAULT_NEW_ALIGNMENT__ as alignment.

+++ DESIRE-Engine/src/Engine/Core/Memory/LinearAllocator.h
@@ -22,8 +22,8 @@ public:
 protected:
 	struct AllocationHeader
 	{
-		uint32_t totalSize;
-		uint32_t offsetToPrev;
+		size_t totalSize;
+		size_t offsetToPrev;
 	};
 
 	bool IsAllocationOwned(const void* ptr) const;
